Added PopUp Class, Updated GUI
A new class has been created for the main popup that the user will first see when opening the project. Currently though it appears along with the slideshow editor, but its intended purpose is to prevent the user from accessing the slideshow editor until an option and its file chooser have been completed.