Updated ClickPopUp and Adding Images to Reel
While currently very janky, adding images/removing images currently works. However this is done by creating a copy of the GUI that the ClickPopUp class has access to. Any changes (adding images) done to the main gui are reflected in the copy of the GUI. I am not sure how to update the main gui with any changes though.